requirements.txt
typer>=0.9.0
rich>=13.0.0
requests>=2.31.0
pyyaml>=6.0.1
pydantic>=2.0.0
psutil>=5.0.0
questionary>=2.0.0
prompt_toolkit>=3.0.0
